refac hamemb
